Update on "[fix] Specify width for st.floats in hypothesis_utils.tensor"
Summary:
circleci complains about generated numbers are not representable by float32 and it pollutes the logs:
https://circleci.com/gh/pytorch/pytorch/2554740?utm_campaign=vcs-integration-link&utm_medium=referral&utm_source=github-build-link

Test Plan:
circleci

Reviewers:
pt1quant
Subscribers:

Tasks:

Tags:

+++ test/hypothesis_utils.py
@@ -40,7 +40,7 @@ def _get_valid_min_max(qparams):
     # make sure intermediate results are within the range of long
     min_value = max((long_min - zero_point) * scale, (long_min / scale + zero_point))
     max_value = min((long_max - zero_point) * scale, (long_max / scale + zero_point))
-    return min_value, max_value
+    return np.float32(min_value), np.float32(max_value)
 
 """Hypothesis filter to avoid overflows with quantized tensors.
